requirements.txt
pyspark==3.5.1
pyyaml
pytest